Embedded the log controller inside of a nav controller.
The nav controller provides a consistent header bar. Moved the segment control to the navigation bar.

Adjusted the leading and trailing constraints of the manual and tracked subviews from -16/16 to -20/20 so that the contoller bg color wasn't showing along the edges of the subviews. This approach should be more solid. Issue #23